Reorganize validator tests
